# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Maiko API Documentation",
        "contact": {
            "email": "juancho637@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/maiko/public/api/v1"
        }
    ],
    "paths": {
...
        "/corrosions/{corrosion}": {
            "get": {
                "tags": [
                    "Corrosiones"
                ],
                "summary": "Muestra la información de una corrosión",
                "operationId": "App\\Http\\Controllers\\Api\\Corrosion\\CorrosionController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de la corrosión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra la información asociada a una corrosión.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Id no encontrado."
                    },
                    "401": {
                        "description": "Usuario no autorizado."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Corrosiones"
                ],
                "summary": "Actualiza una corrosión",
                "operationId": "App\\Http\\Controllers\\Api\\Corrosion\\CorrosionController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de la corrosión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "corrosion_type": {
                                        "type": "string"
                                    },
                                    "remaining_thickness": {
                                        "type": "integer"
                                    },
                                    "area": {
                                        "type": "integer"
                                    },
                                    "large": {
                                        "type": "integer"
                                    },
                                    "thickness": {
                                        "type": "integer"
                                    },
                                    "depth": {
                                        "type": "integer"
                                    },
                                    "observation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualiza una corrosión.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Usuario no autorizado."
                    },
                    "409": {
                        "description": "Error en validaciones de negocio."
                    },
                    "422": {
                        "description": "Entidad no procesable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}